chore: 병합 후 충돌 수정 및 프로필 컴포넌트 분리(#83)
