+++ src/app/(providers)/profile/_components/page.module.scss
@@ -1,3 +1,13 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+  border: 5px solid red;
+  width: 1200px;
+}
+
 .imgLabel {
   display: flex;
   flex-direction: column;
