+++ src/app/(providers)/used-goods/create/page.tsx
@@ -1,4 +1,4 @@
-import CreateForm from './CreateForm';
+import CreateForm from './_components/CreateForm';
 
 const page = () => {
   return (
